avoid deleting a category with products inside

+++ src/app/admin/admin.component.ts
@@ -22,9 +22,6 @@ export class AdminComponent implements OnInit {
 
   ngOnInit(): void {
     this.loadData();
-    this.router.events.subscribe(() => {
-      this.loadData();
-    });
   }
 
   loadData(): void {
@@ -71,7 +68,13 @@ export class AdminComponent implements OnInit {
   }
   
   onDeleteCategory(id: number | undefined) {
+    const productosConCategoriaId = this.products.filter(producto => producto.category === id);
+    if (productosConCategoriaId.length === 0) {
     this.productDataService.deleteCategory(id).subscribe(response => { alert(response.message); window.location.reload(); });
+    }
+    else {
+      alert("La categoria a borrar contiene productos. Intente nuevamente con una vacía.");
+    }
   }
 
 }
